+++ src/dotnet/Fable.Core/Fable.Core.Extensions.fs
@@ -1,18 +1,14 @@
 namespace Fable.Core
 
+open System
+
 [<AutoOpen>]
 module Extensions =
-    open System.Threading
-    open Fable.Core
-    open Fable.Import.JS
-
     type Async with
-        static member AwaitPromise(promise: Promise<'T>): Async<'T> = jsNative
-        static member StartAsPromise(workflow: Async<'T>, ?token: CancellationToken): Promise<'T> = jsNative
+        static member AwaitPromise(promise: JS.Promise<'T>): Async<'T> = jsNative
+        static member StartAsPromise(workflow: Async<'T>, ?token: System.Threading.CancellationToken): JS.Promise<'T> = jsNative
 
 module DynamicExtensions =
-    open System
-
     type Object with
       [<Emit("$0[$1]{{=$2}}")>]
       member __.Item with get(idx: string): obj = jsNative